requirements.txt
pydantic-settings
fastapi==0.70.0
tecton==0.1.0
tortoise-orm==0.17.8
openai==0.11.5
skiff==0.1.0
pandas==1.3.5
